GRHB-386: * renamed default styles file
